requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.2
chardet==3.0.4
idna==2.9
multidict==4.7.6
numpy==1.18.5
python-engineio==3.13.0
python-socketio==4.5.1
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
yarl==1.4.2
